BDA_3_AI_enhanced_ETL/requirements.txt
requests 
beautifulsoup4 
PyPDF2 
python-docx 
pandas